+++ libcaf_core/caf/monitor.test.cpp
@@ -22,28 +22,6 @@ TEST("monitoring another actor") {
   auto client1 = sys.spawn(client_spawn);
   auto client2 = sys.spawn(client_spawn);
   auto client3 = sys.spawn(client_spawn);
-  SECTION("monitoring an actor invokes the down handler") {
-    auto call_count = std::make_shared<int32_t>(0);
-    auto observer = sys.spawn([=](event_based_actor* self) {
-      self->monitor(client1);
-      self->monitor(client2);
-      self->monitor(client3);
-      self->set_down_handler(
-        [call_count](const down_msg&) { *call_count += 1; });
-      return behavior{
-        [](int32_t) {},
-      };
-    });
-    inject_exit(client1);
-    expect<down_msg>().with(std::ignore).to(observer);
-    check_eq(*call_count, 1);
-    inject_exit(client2);
-    expect<down_msg>().with(std::ignore).to(observer);
-    check_eq(*call_count, 2);
-    inject_exit(client3);
-    expect<down_msg>().with(std::ignore).to(observer);
-    check_eq(*call_count, 3);
-  }
   SECTION("monitoring with a callback") {
     auto call_count1 = std::make_shared<int32_t>(0);
     auto call_count2 = std::make_shared<int32_t>(0);
@@ -84,26 +62,24 @@ TEST("monitoring another actor") {
   SECTION("duplicate monitors on one actor deliver two down messages") {
     auto call_count = std::make_shared<int32_t>(0);
     auto observer = sys.spawn([=](event_based_actor* self) {
-      self->monitor(client1);
-      self->monitor(client1);
-      self->set_down_handler(
-        [call_count](const down_msg&) { *call_count += 1; });
+      self->monitor(client1, [call_count](const error&) { *call_count += 1; });
+      self->monitor(client1, [call_count](const error&) { *call_count += 1; });
       return behavior{
         [](int32_t) {},
       };
     });
     inject_exit(client1);
-    expect<down_msg>().with(std::ignore).to(observer);
+    expect<action>().to(observer);
     check_eq(*call_count, 1);
-    expect<down_msg>().with(std::ignore).to(observer);
+    expect<action>().to(observer);
     check_eq(*call_count, 2);
   }
-  SECTION("double demonitor clears duplicate monitors") {
+  SECTION("disposing duplicate monitors clears them") {
     auto observer = sys.spawn([=](event_based_actor* self) {
-      self->monitor(client1);
-      self->monitor(client1);
-      self->demonitor(client1);
-      self->demonitor(client1);
+      auto d1 = self->monitor(client1, [](const error&) {});
+      auto d2 = self->monitor(client1, [](const error&) {});
+      d1.dispose();
+      d2.dispose();
       return behavior{
         [](int32_t) {},
       };
@@ -111,24 +87,26 @@ TEST("monitoring another actor") {
     inject_exit(client1);
     check_eq(mail_count(observer), 0u);
   }
-  SECTION("demonitoring an actor cancels the down handler") {
-    auto call_count = std::make_shared<int32_t>(0);
+  SECTION("disposing a monitor cancels its callback") {
+    auto call_count1 = std::make_shared<int32_t>(0);
+    auto call_count2 = std::make_shared<int32_t>(0);
     auto observer = sys.spawn([=](event_based_actor* self) {
-      self->monitor(client1);
-      self->monitor(client2);
-      self->set_down_handler(
-        [call_count](const down_msg&) { *call_count += 1; });
-      self->demonitor(client1);
+      auto d1 = self->monitor(client1, [call_count1](const error&) {
+        *call_count1 += 1;
+      });
+      self->monitor(client2,
+                    [call_count2](const error&) { *call_count2 += 1; });
+      d1.dispose();
       return behavior{
         [](int32_t) {},
       };
     });
     inject_exit(client1);
     check_eq(mail_count(observer), 0u);
-    check_eq(*call_count, 0);
+    check_eq(*call_count1, 0);
     inject_exit(client2);
-    expect<down_msg>().with(std::ignore).to(observer);
-    check_eq(*call_count, 1);
+    expect<action>().to(observer);
+    check_eq(*call_count2, 1);
   }
   SECTION("canceling a monitor with a callback") {
     auto call_count1 = std::make_shared<int32_t>(0);